pyproject.toml
[project]
name = "llm-automation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "duckdb>=1.2.0",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "markdown>=3.7",
    "numpy>=2.2.3",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "uv>=0.5.31",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
